Closes#24 Removing persistence and vacuum configurable
Unnecessary RDD count to find load time and because of this we are persisting RDD on disk.
This might also create memory issue like disk full if table size is greater than disk.
This also increases time of complete process first loading and then copy the persisted the copied table.